README updated.
Object renamed Geometry.
